# blog-server/docs/swagger.yaml
basePath: /
definitions:
  advert_api.AdvertRequest:
    properties:
      href:
        description: 跳转链接
        type: string
      images:
        description: 图片
        type: string
      is_show:
        description: 是否展示
        type: boolean
      title:
        description: 显示的标题
        type: string
    required:
    - href
    - images
    - title
    type: object
  ctype.ImageType:
    enum:
    - 1
    - 2
    type: integer
    x-enum-comments:
      Local: 本地
      QiNiu: 七牛云
    x-enum-varnames:
    - Local
    - QiNiu
  images_api.ImageResponse:
    properties:
      id:
        description: 图片id
        type: integer
      name:
        description: 图片名称
        type: string
      path:
        description: 图片路径
        type: string
    type: object
  models.AdvertModel:
    properties:
      created_at:
        description: 创建时间
        type: string
      href:
        description: 跳转链接
        type: string
      id:
        description: 主键ID
        type: integer
      images:
        description: 图片
        type: string
      is_show:
        description: 是否展示
        type: boolean
      title:
        description: 显示的标题
        type: string
    type: object
  models.BannerModel:
    properties:
      created_at:
        description: 创建时间
        type: string
      hash:
        description: 图片的hash值，用于判断重复图片
        type: string
      id:
        description: 主键ID
        type: integer
      image_type:
        allOf:
        - $ref: '#/definitions/ctype.ImageType'
        description: 图片的类型， 本地还是七牛
      name:
        description: 图片名称
        type: string
      path:
        description: 图片路径
        type: string
    type: object
  models.RemoveRequest:
    properties:
      id_list:
        items:
          type: integer
        type: array
    type: object
  res.ListResponse-models_AdvertModel:
    properties:
      count:
        type: integer
      list:
        $ref: '#/definitions/models.AdvertModel'
    type: object
  res.ListResponse-models_BannerModel:
    properties:
      count:
        type: integer
      list:
        $ref: '#/definitions/models.BannerModel'
    type: object
  res.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: server项目API文档
  title: server API文档
  version: "1"
paths:
  /api/adverts:
    delete:
      description: 删除广告
      parameters:
      - description: 广告id列表
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.RemoveRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/res.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: 删除广告
      tags:
      - 广告管理
    get:
      description: 广告列表
      parameters:
      - description: 查询关键字
        in: query
        name: key
        type: string
      - description: 每页条数
        in: query
        name: limit
        type: integer
      - description: 当前页
        in: query
        name: page
        type: integer
      - in: query
        name: sort
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/res.Response'
            - properties:
                data:
                  $ref: '#/definitions/res.ListResponse-models_AdvertModel'
              type: object
      summary: 广告列表
      tags:
      - 广告管理
    post:
      description: 创建广告
      parameters:
      - description: 广告信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/advert_api.AdvertRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/res.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/res.Response'
      summary: 创建广告
      tags:
      - 广告管理
  /api/adverts/:id:
    put:
      description: 更新广告
      parameters:
      - description: 广告更新信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/advert_api.AdvertRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/res.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: 更新广告
      tags:
      - 广告管理
  /api/images:
    get:
      description: 图片列表
      parameters:
      - description: 查询关键字
        in: query
        name: key
        type: string
      - description: 每页条数
        in: query
        name: limit
        type: integer
      - description: 当前页
        in: query
        name: page
        type: integer
      - in: query
        name: sort
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/res.Response'
            - properties:
                data:
                  $ref: '#/definitions/res.ListResponse-models_BannerModel'
              type: object
      summary: 图片列表
      tags:
      - 图片管理
  /api/images_names:
    get:
      description: 图片名称列表
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/res.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/images_api.ImageResponse'
                  type: array
              type: object
      summary: 图片名称列表
      tags:
      - 图片管理
swagger: "2.0"